+++ src/main/java/org/thema/pixscape/LandModTask.java
@@ -38,6 +38,7 @@ import org.geotools.feature.SchemaException;
 import org.thema.common.collection.HashMapList;
 import org.thema.data.IOImage;
 import org.thema.data.feature.DefaultFeature;
+import org.thema.data.IOFeature;
 import org.thema.parallel.AbstractParallelTask;
 
 /**
@@ -91,7 +92,7 @@ public class LandModTask extends AbstractParallelTask<Void, Void> implements Ser
                 project = Project.load(prjFile);
             }
         
-            List<DefaultFeature> features = DefaultFeature.loadFeatures(fileZone);
+            List<DefaultFeature> features = IOFeature.loadFeatures(fileZone);
             if(!features.get(0).getAttributeNames().contains(idField)) {
                 throw new IllegalArgumentException("Unknow field : " + idField);
             }